services/web/appengine/requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.4
aniso8601==9.0.1
black==22.10.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.42
tomli==2.0.1
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.9.0
celery==5.2.3
redis==4.1.4
pg8000==1.29.3
cloud-sql-python-connector==0.9.0
SQLAlchemy==1.4.42
google-cloud-storage==2.6.0
google-cloud-pubsub==2.13.11
